Update README.md
A key innovation in this project is the implementation of real-time data sharing between the fertilizer robot and the
harvesting robot. Real-time data sharing allows for seamless communication and coordination between the two robots,
enhancing their operational efficiency and effectiveness. For instance, if the fertilizer robot detects a diseased plant, it
can mark its location and transmit this information to the harvesting robot, enabling it to avoid the affected area or take
specific actions to manage the disease. The harvesting robot is equipped with advanced navigation systems, cameras,
and a harvesting mechanism that can identify and pick ripe chilies accurately. The integration of real-time data sharing
ensures that the harvesting process is optimized, reducing crop loss and improving overall productivity.